\name{quantlab2-package}
\alias{quantlab2-package}
\alias{quantlab2}
\docType{package}
\title{Functions applied to options pricing}
\description{A collection of functions to valuate options, greeks, volatility, and other parameters used by traders. Vanna-Volga, realized and implied volatility, skewness, and kurtosis, volatility smile scenarios, and other techniques are included}
\details{
\tabular{ll}{
Package: \tab quantlab2\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-11-30\cr
License: \tab GPL (>= 2)\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Oleg Mubarakshin
Maintainer: Oleg Mubarakshin <oleg.mubarakshin@gmail.com>
}
\references{\url{quant-lab.com}}
\keyword{vanna-volga}
\examples{}
