\name{quantlab2-package}
\alias{quantlab2-package}
\alias{quantlab2}
\docType{package}
\title{
Options Trading Framework
}
\description{
 A collection of function to valuate options, greeks, volatility, and other parameters used by traders. Vanna-Volga, realized and implied volatility, skewness, and kurtosis, volatility smile scenarios, P&L profiles, and other techniques are included
}
\details{
\tabular{ll}{
Package: \tab quantlab2\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2014-05-01\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Oleg Mubarakshin

Maintainer: Oleg Mubarakshin <oleg.mubarakshin@gmail.com>
}
\references{
1. Mubarakshin O., 2014, \href{http://quant-lab.com/wp-content/uploads/2014/03/Oleg-Mubarakshin-Quant-lab.pdf/}{"Quant-lab: Options Trading & Market Making Tools"}, paper for the Sixth Annual R/Finance Conference for applied finance using R (May 16 and 17, 2014 in Chicago, IL, USA at the University of Illinois at Chicago), January 27, 2014

2. Castagna A., Mercurio F, 2007, "The vanna-volga method for implied volatilities", Risk Magazine, January 2007, 106-111 

\url{quant-lab.com}

\href{http://quant-lab.r-forge.r-project.org/}{quant-lab on the R-Forge}

}

%%\keyword{}

%%\seealso{}

%%\examples{}
